# frontend/contracts/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sadhana Card Tracker API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/registrations/pending": {
      "get": {
        "tags": [
          "admin-users"
        ],
        "summary": "Pending Registrations",
        "operationId": "pending_registrations_api_v1_admin_registrations_pending_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "type": "array",
                  "title": "Response Pending Registrations Api V1 Admin Registrations Pending Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/registrations/{user_id}/approve": {
      "post": {
        "tags": [
          "admin-users"
        ],
        "summary": "Approve",
        "operationId": "approve_api_v1_admin_registrations__user_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/registrations/{user_id}/reject": {
      "post": {
        "tags": [
          "admin-users"
        ],
        "summary": "Reject",
        "operationId": "reject_api_v1_admin_registrations__user_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/devotees": {
      "post": {
        "tags": [
          "admin-users"
        ],
        "summary": "Admin Create Devotee",
        "operationId": "admin_create_devotee_api_v1_admin_devotees_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCreateDevoteeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin-users"
        ],
        "summary": "Devotees",
        "operationId": "devotees_api_v1_admin_devotees_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "account_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AccountStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserPublic"
                  },
                  "title": "Response Devotees Api V1 Admin Devotees Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/devotees/{user_id}/deactivate": {
      "post": {
        "tags": [
          "admin-users"
        ],
        "summary": "Deactivate",
        "operationId": "deactivate_api_v1_admin_devotees__user_id__deactivate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/devotees/{user_id}/activate": {
      "post": {
        "tags": [
          "admin-users"
        ],
        "summary": "Activate",
        "operationId": "activate_api_v1_admin_devotees__user_id__activate_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountStatusChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccessTokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "AccessTokenResponse"
      },
      "AccountStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "ACTIVE",
          "INACTIVE",
          "REJECTED"
        ],
        "title": "AccountStatus"
      },
      "AccountStatusChangeRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "AccountStatusChangeRequest"
      },
...
      "AdminCreateDevoteeRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "phone_number": {
            "type": "string",
            "maxLength": 30,
            "minLength": 7,
            "title": "Phone Number"
          },
          "college": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "College"
          },
          "branch": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Branch"
          },
          "college_joining_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "College Joining Year"
          },
          "category_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Category Code"
          },
          "current_academic_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 4.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Academic Year"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "password",
          "phone_number",
          "college",
          "branch",
          "college_joining_year",
          "category_code"
        ],
        "title": "AdminCreateDevoteeRequest"
      },
...
      "CategorySummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "academic_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Academic Year"
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "display_name",
          "academic_year"
        ],
        "title": "CategorySummary"
      },
...
      "DevoteeProfilePublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "college": {
            "type": "string",
            "title": "College"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "college_joining_year": {
            "type": "integer",
            "title": "College Joining Year"
          },
          "expected_graduation_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Graduation Year"
          },
          "current_academic_year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Academic Year"
          },
          "current_category": {
            "$ref": "#/components/schemas/CategorySummary"
          }
        },
        "type": "object",
        "required": [
          "id",
          "college",
          "branch",
          "college_joining_year",
          "expected_graduation_year",
          "current_academic_year",
          "current_category"
        ],
        "title": "DevoteeProfilePublic"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LogoutResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "LogoutResponse"
      },
...
      "RegistrationRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          },
          "phone_number": {
            "type": "string",
            "maxLength": 30,
            "minLength": 7,
            "title": "Phone Number"
          },
          "college": {
            "type": "string",
            "maxLength": 200,
            "minLength": 2,
            "title": "College"
          },
          "branch": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Branch"
          },
          "current_academic_year": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Current Academic Year"
          },
          "college_joining_year": {
            "type": "integer",
            "maximum": 2100.0,
            "minimum": 2000.0,
            "title": "College Joining Year"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "email",
          "password",
          "phone_number",
          "college",
          "branch",
          "current_academic_year",
          "college_joining_year"
        ],
        "title": "RegistrationRequest"
      },
      "RegistrationResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "account_status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "email",
          "account_status",
          "message"
        ],
        "title": "RegistrationResponse"
      },
      "RegistrationSource": {
        "type": "string",
        "enum": [
          "SELF_REGISTERED",
          "ADMIN_CREATED"
        ],
        "title": "RegistrationSource"
      },
      "RejectionRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "title": "RejectionRequest"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone_number": {
            "type": "string",
            "title": "Phone Number"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "account_status": {
            "$ref": "#/components/schemas/AccountStatus"
          },
          "registration_source": {
            "$ref": "#/components/schemas/RegistrationSource"
          },
          "approved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "deactivated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deactivated At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "devotee_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DevoteeProfilePublic"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "email",
          "phone_number",
          "role",
          "account_status",
          "registration_source",
          "approved_at",
          "deactivated_at",
          "created_at",
          "devotee_profile"
        ],
        "title": "UserPublic"
      },
      "UserRole": {
        "type": "string",
        "enum": [
          "DEVOTEE",
          "ADMIN"
        ],
        "title": "UserRole"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}